delete request confirm

+++ javascript/updateproducts/deleteproduct.js
@@ -11,22 +11,27 @@ export function deleteProduct(id) {
   deleteButton.onclick = async function () {
     console.log(id);
 
-    const deleteUrl = baseUrl + "products/" + id;
-    const token = fetchToken();
+    const deleteWarning = confirm("are you sure you want to delete this product?");
+    console.log(deleteWarning);
+    if (deleteWarning) {
+      const deleteUrl = baseUrl + "products/" + id;
+      const token = fetchToken();
 
-    const delOptions = {
-      method: "DELETE",
-      headers: {
-        "Content-type": "application/json",
-        Authorization: `Bearer ${token}`,
-      },
-    };
-    try {
-      const response = await fetch(deleteUrl, delOptions);
-      const results = await response.json();
-      console.log(results);
-    } catch (error) {
-      console.log(error.message);
+      const delOptions = {
+        method: "DELETE",
+        headers: {
+          "Content-type": "application/json",
+          Authorization: `Bearer ${token}`,
+        },
+      };
+      try {
+        const response = await fetch(deleteUrl, delOptions);
+        const results = await response.json();
+        console.log(results);
+        location.href = "/editproduct.html";
+      } catch (error) {
+        console.log(error.message);
+      }
     }
   };
 }
